Player

See merge request TDDC76-2022-PRA1/Nils-2!33

+++ player.h
@@ -20,13 +20,16 @@ public:
     void collision(Game_object* object) override;
     bool check_collision(Game_object* object) override;
 
+    
+
 
 protected:
     int hp{};
 
 private:
-    int bearing{};
+    double bearing{};
     int score{};
+    
 };
